--- a/package/yast2-users.changes
+++ b/package/yast2-users.changes
@@ -1,4 +1,10 @@
 -------------------------------------------------------------------
+Wed Sep 21 13:25:22 UTC 2016 - mvidner@suse.com
+
+- Do not require yast2-ldap for build time tests (bsc#999203).
+- 3.1.59
+
+-------------------------------------------------------------------
 Fri Sep 16 09:03:32 UTC 2016 - igonzalezsosa@suse.com
 
 - Add support to specify SSH authorized keys in the profile