--- a/package/yast2-users.spec
+++ b/package/yast2-users.spec
@@ -17,7 +17,7 @@
 
 
 Name:           yast2-users
-Version:        3.1.58
+Version:        3.1.59
 Release:        0
 
 BuildRoot:      %{_tmppath}/%{name}-%{version}-build