[bsc#999203](https://bugzilla.suse.com/show_bug.cgi?id=999203)

Most importantly, this breaks the critical path link between yast2-users and yast2-network (users no longer BuildRequire ldap which Requires network)

Total build time for this package on my machine decreases from 69s to 63s

The number of buildrequired packages decreases from 255 to 227:

```
[    2s] deleting curl
[    2s] deleting device-mapper
[    2s] deleting elfutils
[    2s] deleting libaio1
[    2s] deleting libasm1
[    2s] deleting libdevmapper-event1_03
[    2s] deleting libdevmapper1_03
[    2s] deleting libdw1
[    2s] deleting libldapcpp1
[    2s] deleting libmetalink3
[    2s] deleting libparted0
[    3s] deleting libstorage-ruby
[    3s] deleting libstorage7
[    3s] deleting lsscsi
[    3s] deleting parted
[    3s] deleting ruby2.2-rubygem-ruby-dbus
[    3s] deleting suse-module-tools
[    3s] deleting thin-provisioning-tools
[    3s] deleting udev-mini
[    3s] deleting unzip
[    3s] deleting yast2-country-data
[    3s] deleting yast2-ldap
[    3s] deleting yast2-network
[    3s] deleting yast2-packager
[    4s] deleting yast2-proxy
[    4s] deleting yast2-storage
[    4s] deleting yast2-transfer
[    4s] deleting yast2-xml
```