@mvidner 2016-09-21
Ugh, Travis behaves differently.